+++ packages/stream_chat_persistence/lib/src/entity/poll_votes.dart
@@ -1,5 +1,6 @@
 // coverage:ignore-file
 import 'package:drift/drift.dart';
+import '../converter/map_converter.dart';
 import 'entity.dart';
 
 /// Represents a [PollVotes] table in [DriftChatDatabase].
@@ -20,6 +21,9 @@ class PollVotes extends Table {
   ///
   /// Nullable if the user selected an option.
   TextColumn get answerText => text().nullable()();
+
+  /// A Map of [answerText] translations.
+  TextColumn get answerTextI18n => text().nullable().map(NullableMapConverter<String>())();
 
   /// The date when the poll vote was created.
   DateTimeColumn get createdAt => dateTime().withDefault(currentDateAndTime)();